+++ backend/chainlit/__init__.py
@@ -20,16 +20,18 @@ if TYPE_CHECKING:
     )
     from chainlit.llama_index.callbacks import LlamaIndexCallbackHandler
     from chainlit.openai import instrument_openai
+    from chainlit.mistralai import instrument_mistralai
 
 import chainlit.input_widget as input_widget
 from chainlit.action import Action
 from chainlit.cache import cache
+from chainlit.chat_context import chat_context
 from chainlit.chat_settings import ChatSettings
 from chainlit.config import config
 from chainlit.context import context
 from chainlit.element import (
     Audio,
-    Avatar,
+    Component,
     File,
     Image,
     Pdf,
@@ -53,7 +55,7 @@ from chainlit.oauth_providers import get_configured_oauth_providers
 from chainlit.step import Step, step
 from chainlit.sync import make_async, run_sync
 from chainlit.telemetry import trace
-from chainlit.types import AudioChunk, ChatProfile, ThreadDict
+from chainlit.types import AudioChunk, ChatProfile, Starter, ThreadDict
 from chainlit.user import PersistedUser, User
 from chainlit.user_session import user_session
 from chainlit.utils import make_module_getattr, wrap_user_function
@@ -116,7 +118,7 @@ def oauth_callback(
 
     Example:
         @cl.oauth_callback
-        async def oauth_callback(provider_id: str, token: str, raw_user_data: Dict[str, str], default_app_user: User) -> Optional[User]:
+        async def oauth_callback(provider_id: str, token: str, raw_user_data: Dict[str, str], default_app_user: User, id_token: Optional[str]) -> Optional[User]:
 
     Returns:
         Callable[[str, str, Dict[str, str], User], Optional[User]]: The decorated authentication callback.
@@ -206,6 +208,22 @@ def set_chat_profiles(
     """
 
     config.code.set_chat_profiles = wrap_user_function(func)
+    return func
+
+
+@trace
+def set_starters(func: Callable[[Optional["User"]], List["Starter"]]) -> Callable:
+    """
+    Programmatic declaration of the available starter (can depend on the User from the session if authentication is setup).
+
+    Args:
+        func (Callable[[Optional["User"]], List["Starter"]]): The function declaring the starters.
+
+    Returns:
+        Callable[[Optional["User"]], List["Starter"]]: The decorated function.
+    """
+
+    config.code.set_starters = wrap_user_function(func)
     return func
 
 
@@ -346,11 +364,15 @@ __getattr__ = make_module_getattr(
         "LlamaIndexCallbackHandler": "chainlit.llama_index.callbacks",
         "HaystackAgentCallbackHandler": "chainlit.haystack.callbacks",
         "instrument_openai": "chainlit.openai",
+        "instrument_mistralai": "chainlit.mistralai",
     }
 )
 
 __all__ = [
+    "ChatProfile",
+    "Starter",
     "user_session",
+    "chat_context",
     "CopilotFunction",
     "AudioChunk",
     "Action",
@@ -361,7 +383,7 @@ __all__ = [
     "Plotly",
     "Image",
     "Text",
-    "Avatar",
+    "Component",
     "Pyplot",
     "File",
     "Task",
@@ -401,6 +423,7 @@ __all__ = [
     "LlamaIndexCallbackHandler",
     "HaystackAgentCallbackHandler",
     "instrument_openai",
+    "instrument_mistralai",
 ]