+++ frontend/src/components/atoms/logo.tsx
@@ -1,7 +1,9 @@
-import { apiClient } from 'api';
+import { useContext } from 'react';
 import { useRecoilValue } from 'recoil';
 
 import { settingsState } from 'state/settings';
+
+import { ChainlitContext } from 'client-types/*';
 
 interface Props {
   width?: number;
@@ -10,6 +12,7 @@ interface Props {
 
 export const Logo = ({ style }: Props) => {
   const { theme } = useRecoilValue(settingsState);
+  const apiClient = useContext(ChainlitContext);
 
   return (
     <img src={apiClient.getLogoEndpoint(theme)} alt="logo" style={style} />