+++ frontend/src/components/molecules/messages/Message.tsx
@@ -1,13 +1,12 @@
 import { keyframes } from '@emotion/react';
 import { MessageContext } from 'contexts/MessageContext';
-import { memo, useContext, useEffect, useState } from 'react';
+import { memo, useContext } from 'react';
 
 import Box from '@mui/material/Box';
 import Stack from '@mui/material/Stack';
 
 import { AskUploadButton } from './components/AskUploadButton';
-import { Author } from './components/Author';
-import { DetailsButton } from './components/DetailsButton';
+import { MessageAvatar } from './components/Avatar';
 import { MessageActions } from './components/MessageActions';
 import { MessageButtons } from './components/MessageButtons';
 import { MessageContent } from './components/MessageContent';
@@ -17,33 +16,22 @@ import { useLayoutMaxWidth } from 'hooks/useLayoutMaxWidth';
 import type { IAction, IMessageElement, IStep } from 'client-types/';
 
 import BlinkingCursor from '../BlinkingCursor';
-import { Messages } from './Messages';
+import ToolCalls from './ToolCalls';
+import UserMessage from './UserMessage';
 
 interface Props {
   message: IStep;
+  showAvatar?: boolean;
   elements: IMessageElement[];
   actions: IAction[];
   indent: number;
-  showAvatar?: boolean;
-  showBorder?: boolean;
   isRunning?: boolean;
   isLast?: boolean;
 }
 
 const Message = memo(
-  ({
-    message,
-    elements,
-    actions,
-    indent,
-    showAvatar,
-    showBorder,
-    isRunning,
-    isLast
-  }: Props) => {
+  ({ message, showAvatar, elements, actions, isRunning, isLast }: Props) => {
     const {
-      expandAll,
-      hideCot,
       highlightedMessage,
       defaultCollapseContent,
       allowHtml,
@@ -51,21 +39,11 @@ const Message = memo(
       onError
     } = useContext(MessageContext);
     const layoutMaxWidth = useLayoutMaxWidth();
-
-    const [showDetails, setShowDetails] = useState(expandAll);
-
-    useEffect(() => {
-      setShowDetails(expandAll);
-    }, [expandAll]);
-
-    if (hideCot && indent) {
-      return null;
-    }
-
     const isAsk = message.waitForAnswer;
 
     const isUserMessage = message.type === 'user_message';
 
+    const forceDisplayCursor = isLast && isRunning && !message.streaming;
     return (
       <Box
         sx={{
@@ -89,9 +67,7 @@ const Message = memo(
             id={`step-${message.id}`}
             direction="row"
             sx={{
-              py: 2,
-              borderBottom: (theme) =>
-                showBorder ? `1px solid ${theme.palette.divider}` : 'none',
+              pb: 2,
               animation:
                 message.id && highlightedMessage === message.id
                   ? `3s ease-in-out 0.1s ${flash}`
@@ -101,16 +77,7 @@ const Message = memo(
           >
             {isUserMessage ? (
               <Box display="flex" flexDirection="column" width="100%">
-                <Box
-                  sx={{
-                    px: 2.5,
-                    py: 1,
-                    borderRadius: '1.5rem',
-                    backgroundColor: 'background.paper',
-                    maxWidth: '90%',
-                    ml: 'auto'
-                  }}
-                >
+                <UserMessage message={message}>
                   <MessageContent
                     elements={elements}
                     message={message}
@@ -120,41 +87,39 @@ const Message = memo(
                     allowHtml={allowHtml}
                     latex={latex}
                   />
-                </Box>
-                <DetailsButton
-                  message={message}
-                  opened={showDetails}
-                  onClick={() => setShowDetails(!showDetails)}
-                  loading={isRunning && isLast}
-                />
-                {message.steps && showDetails && (
-                  <Messages
-                    messages={message.steps}
-                    actions={actions}
-                    elements={elements}
-                    indent={indent + 1}
-                    isRunning={isRunning}
-                  />
+                </UserMessage>
+                {forceDisplayCursor && (
+                  <Stack
+                    direction="row"
+                    gap="1rem"
+                    alignItems="center"
+                    my={2}
+                    width="100%"
+                  >
+                    <MessageAvatar />
+                    <BlinkingCursor />
+                  </Stack>
                 )}
               </Box>
             ) : (
-              <Author message={message} show={showAvatar}>
-                <DetailsButton
-                  message={message}
-                  opened={showDetails}
-                  onClick={() => setShowDetails(!showDetails)}
-                  loading={isRunning && isLast}
-                />
-                {message.steps && showDetails && (
-                  <Messages
-                    messages={message.steps}
-                    actions={actions}
+              <Stack
+                direction="row"
+                gap="1rem"
+                width="100%"
+                className="ai-message"
+              >
+                <MessageAvatar author={message.name} hide={!showAvatar} />
+                <Stack
+                  alignItems="flex-start"
+                  minWidth={150}
+                  flexGrow={1}
+                  position="relative"
+                >
+                  <ToolCalls
                     elements={elements}
-                    indent={indent + 1}
+                    message={message}
                     isRunning={isRunning}
                   />
-                )}
-                <Stack alignItems="flex-start" minWidth={150}>
                   <MessageContent
                     elements={elements}
                     message={message}
@@ -171,29 +136,15 @@ const Message = memo(
                     <MessageActions message={message} actions={actions} />
                   ) : null}
                   <MessageButtons message={message} />
+                  {forceDisplayCursor && (
+                    <Box my={0.5}>
+                      <BlinkingCursor />
+                    </Box>
+                  )}
                 </Stack>
-              </Author>
+              </Stack>
             )}
           </Stack>
-          {isLast &&
-            isRunning &&
-            (!message.streaming || window.renderingCodeBlock) && (
-              <Box
-                sx={{
-                  position: 'absolute',
-                  bottom: -5,
-                  left: 38,
-                  boxSizing: 'border-box',
-                  mx: 'auto',
-                  maxWidth: layoutMaxWidth,
-                  px: 2,
-                  display: 'flex',
-                  flexDirection: 'column'
-                }}
-              >
-                <BlinkingCursor />
-              </Box>
-            )}
         </Box>
       </Box>
     );