+++ frontend/src/components/molecules/messages/components/MessageButtons.tsx
@@ -4,14 +4,16 @@ import { grey } from 'theme/palette';
 
 import Stack from '@mui/material/Stack';
 
+import { useChatMessages, useConfig } from '@chainlit/react-client';
+
 import { ClipboardCopy } from 'components/atoms/ClipboardCopy';
 
 import { useIsDarkMode } from 'hooks/useIsDarkMode';
 
-import type { IStep } from 'client-types/';
+import { type IStep } from 'client-types/';
 
+import { DebugButton } from './DebugButton';
 import { FeedbackButtons } from './FeedbackButtons';
-import { PlaygroundButton } from './PlaygroundButton';
 
 interface Props {
   message: IStep;
@@ -20,8 +22,9 @@ interface Props {
 const MessageButtons = ({ message }: Props) => {
   const isDark = useIsDarkMode();
   const { showFeedbackButtons: showFbButtons } = useContext(MessageContext);
+  const { config } = useConfig();
+  const { firstInteraction } = useChatMessages();
 
-  const showPlaygroundButton = !!message.generation;
   const isUser = message.type === 'user_message';
   const isAsk = message.waitForAnswer;
   const hasContent = !!message.output;
@@ -35,22 +38,27 @@ const MessageButtons = ({ message }: Props) => {
     !isAsk &&
     hasContent;
 
-  const show = showCopyButton || showPlaygroundButton || showFeedbackButtons;
+  const showDebugButton =
+    !!config?.debugUrl && !!message.threadId && !!firstInteraction;
 
-  if (!show) {
+  const show = showCopyButton || showDebugButton || showFeedbackButtons;
+
+  if (!show || message.streaming) {
     return null;
   }
 
   return (
     <Stack
-      sx={{ marginLeft: '-8px !important', mt: 0.5 }}
+      sx={{ marginLeft: '-8px !important' }}
       alignItems="center"
       direction="row"
       color={isDark ? grey[400] : grey[600]}
     >
       {showCopyButton ? <ClipboardCopy value={message.output} /> : null}
       {showFeedbackButtons ? <FeedbackButtons message={message} /> : null}
-      {showPlaygroundButton ? <PlaygroundButton step={message} /> : null}
+      {showDebugButton ? (
+        <DebugButton debugUrl={config.debugUrl!} step={message} />
+      ) : null}
     </Stack>
   );
 };