+++ frontend/src/components/molecules/messages/components/MessageContent.tsx
@@ -29,9 +29,10 @@ const MessageContent = memo(
     let lineCount = 0;
     let contentLength = 0;
 
-    const content = message.streaming
-      ? message.output + CURSOR_PLACEHOLDER
-      : message.output;
+    const outputContent =
+      message.streaming && message.output
+        ? message.output + CURSOR_PLACEHOLDER
+        : message.output;
 
     const {
       preparedContent: output,
@@ -40,7 +41,7 @@ const MessageContent = memo(
     } = prepareContent({
       elements,
       id: message.id,
-      content: content,
+      content: outputContent,
       language: message.language
     });
 
@@ -60,11 +61,15 @@ const MessageContent = memo(
     let inputMarkdown;
 
     if (message.input && message.showInput) {
+      const inputContent =
+        message.streaming && message.input
+          ? message.input + CURSOR_PLACEHOLDER
+          : message.input;
       const { preparedContent: input, refElements: inputRefElements } =
         prepareContent({
           elements,
           id: message.id,
-          content: message.input,
+          content: inputContent,
           language:
             typeof message.showInput === 'string'
               ? message.showInput
@@ -103,7 +108,6 @@ const MessageContent = memo(
 
     const collapse =
       lineCount > COLLAPSE_MIN_LINES || contentLength > COLLAPSE_MIN_LENGTH;
-
     const messageContent = collapse ? (
       <Collapse defaultExpandAll={preserveSize}>{markdownContent}</Collapse>
     ) : (
@@ -111,9 +115,9 @@ const MessageContent = memo(
     );
 
     return (
-      <Stack width="100%" direction="row">
+      <Stack width="100%" direction="row" className="message-content">
         <Box width="100%">
-          {output ? messageContent : null}
+          {!!inputMarkdown || output ? messageContent : null}
           <InlinedElements elements={outputInlinedElements} />
         </Box>
       </Stack>