+++ frontend/src/components/molecules/tasklist/TaskList.tsx
@@ -1,4 +1,4 @@
-import { useFetch } from 'usehooks-ts';
+import { useMemo } from 'react';
 import { grey } from 'theme';
 
 import { Box, Chip, List, Theme, useTheme } from '@mui/material';
@@ -21,7 +21,12 @@ const Header = ({ status }: { status: string }) => {
       }}
     >
       <Box
-        sx={{ flexGrow: '1', fontWeight: '600', paddingLeft: theme.spacing(1) }}
+        sx={{
+          flexGrow: '1',
+          fontWeight: '600',
+          paddingLeft: theme.spacing(1),
+          fontFamily: theme.typography.fontFamily
+        }}
       >
         <Translator path="components.molecules.tasklist.TaskList.title" />
       </Box>
@@ -44,6 +49,7 @@ const taskListContainerStyles = (theme: Theme) => ({
   width: '100%',
   display: 'flex',
   flexDirection: 'column',
+  fontFamily: theme.typography.fontFamily!,
   boxShadow:
     theme.palette.mode === 'dark'
       ? '0px 4px 20px 0px rgba(0, 0, 0, 0.20)'
@@ -56,13 +62,21 @@ const TaskList = ({ isMobile }: { isMobile: boolean }) => {
 
   const tasklist = tasklists[tasklists.length - 1];
 
-  const url = tasklist?.url;
+  // We remove the base URL since the useApi hook is already set with a base URL.
+  // This ensures we only pass the relative path and search parameters to the hook.
+  const url = useMemo(() => {
+    if (!tasklist?.url) return null;
+    const parsedUrl = new URL(tasklist.url);
+    return parsedUrl.pathname + parsedUrl.search;
+  }, [tasklist?.url]);
 
-  const { data, error } = useFetch(url);
+  const { isLoading, error, data } = useApi<ITaskList>(url ? url : null, {
+    keepPreviousData: true
+  });
 
   if (!url) return null;
 
-  if (!data && !error) {
+  if (isLoading && !data) {
     return (
       <div>
         <Translator path="components.molecules.tasklist.TaskList.loading" />