+++ frontend/src/contexts/MessageContext.tsx
@@ -4,14 +4,12 @@ import { IMessageContext } from 'types/messageContext';
 
 const defaultMessageContext = {
   avatars: [],
-  defaultCollapseContent: true,
-  expandAll: false,
+  defaultCollapseContent: false,
   hideCot: false,
   highlightedMessage: null,
   loading: false,
   onElementRefClick: undefined,
   onFeedbackUpdated: undefined,
-  onPlaygroundButtonClick: undefined,
   showFeedbackButtons: true,
   onError: () => undefined,
   uiName: ''