+++ frontend/src/pages/Login.tsx
@@ -1,7 +1,4 @@
-import { apiClient } from 'api';
-import { useAuth } from 'api/auth';
-import { useEffect, useState } from 'react';
-import { useTranslation } from 'react-i18next';
+import { useContext, useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 
 import { Logo } from 'components/atoms/logo';
@@ -10,14 +7,16 @@ import { AuthLogin } from 'components/molecules/auth';
 
 import { useQuery } from 'hooks/query';
 
+import { ChainlitContext, useAuth } from 'client-types/*';
+
 export default function Login() {
   const query = useQuery();
   const { data: config, setAccessToken, user } = useAuth();
   const [error, setError] = useState('');
+  const apiClient = useContext(ChainlitContext);
 
   const navigate = useNavigate();
 
-  const { t } = useTranslation();
   const handleHeaderAuth = async () => {
     try {
       const json = await apiClient.headerAuth();
@@ -67,7 +66,7 @@ export default function Login() {
 
   return (
     <AuthLogin
-      title="Evidnet Chatbot"
+      title={<Translator path="components.molecules.auth.authLogin.title" />}
       error={error}
       callbackUrl="/"
       providers={config?.oauthProviders || []}