+++ frontend/src/pages/Thread.tsx
@@ -1,35 +1,26 @@
 import { useEffect } from 'react';
 import { useParams } from 'react-router-dom';
-import { useRecoilState, useRecoilValue } from 'recoil';
+import { useRecoilState } from 'recoil';
 
 import { Box } from '@mui/material';
 
 import {
-  accessTokenState,
   IThread,
   threadHistoryState,
   useApi,
   useChatMessages
 } from '@chainlit/react-client';
-import { threadsFiltersState } from "../state/threads";
 
 import Chat from 'components/organisms/chat';
 import { Thread } from 'components/organisms/sidebar/threadHistory/Thread';
 
-import { apiClientState } from 'state/apiClient';
-
 import Page from './Page';
 import ResumeButton from './ResumeButton';
-import { fetchThreads } from 'state/fetchThreads';
-
 
 export default function ThreadPage() {
   const { id } = useParams();
-  const apiClient = useRecoilValue(apiClientState);
-  const accessToken = useRecoilValue(accessTokenState);
-  const filters = useRecoilValue(threadsFiltersState);
+
   const { data, error, isLoading } = useApi<IThread>(
-    apiClient,
     id ? `/project/thread/${id}` : null,
     {
       revalidateOnFocus: false
@@ -41,12 +32,6 @@ export default function ThreadPage() {
   const { threadId } = useChatMessages();
 
   const isCurrentThread = threadId === id;
-
-  const handleFetchThreads = async () => {
-    if (threadHistory) { // threadHistory가 undefined가 아닌 경우에만 fetchThreads 호출
-      await fetchThreads(threadHistory, setThreadHistory, accessToken, filters);
-    }
-  };
 
   useEffect(() => {
     if (threadHistory?.currentThreadId !== id) {
@@ -69,7 +54,10 @@ export default function ThreadPage() {
               gap: 2
             }}
           >
-            <ResumeButton threadId={id} isLoading={isLoading} onThreadsFetched={handleFetchThreads} />
+            <Box sx={{ width: '100%', flexGrow: 1, overflow: 'auto' }}>
+              <Thread thread={data} error={error} isLoading={isLoading} />
+            </Box>
+            <ResumeButton threadId={id} />
           </Box>
         )}
       </>